feat: implement logging and token validation services for MCP server management
